+++ modules/Users/User.php
@@ -334,6 +334,8 @@ if ( $_REQUEST['modfunc'] === 'update'
 				{
 					DBQuery( $sql );
 
+					$note[] = button( 'check' ) . ' ' . _( 'Your changes were saved.' );
+
 					// Hook.
 					do_action( 'Users/User.php|update_user' );
 				}