+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 6.5
 - Disable save, search & sort for Notification Results list in NotifiyParents.php
 - Fix empty template saved in CreateParents.php & NotifiyParents.php
 - Fix SQL error when no User ID returned by Moodle in Moodle/Custom/CreateParents.php
+- Do not allow Parents/Students to delete Existing Address/Contact in Address.inc.php
 
 Changes in 6.4.2
 ----------------