+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 6.5
 - Fix SQL error when no User ID returned by Moodle in Moodle/Custom/CreateParents.php
 - Do not allow Parents/Students to delete Existing Address/Contact in Address.inc.php
 - Add Profile to ErrorSendEmail() in ErrorMessage.fnc.php
+- Course Widget: add Subject and Not options in Widgets.fnc.php, Scheduling/functions.inc.php & ChooseCourse.php, sponsored by English National Program
 
 Changes in 6.4.2
 ----------------